update github workflow

+++ install.sh
@@ -3,11 +3,12 @@ echo "Installing dependencies and activating virtualenv..."
 
 set -e
 
-rm -rf .venv/
-
-python3.10 -m venv .venv
-
-source .venv/bin/activate
+if [[ "$SKIP_VIRTUAL_ENV" -eq 0 ]]; then
+    echo "(re)setting virtual environment"
+    rm -rf .venv/
+    python3.10 -m venv .venv
+    source .venv/bin/activate
+fi
 
 pip install --upgrade pip
 pip3 install -r requirements.txt
